Move platform-specific logic to a single file
