+++ minijvm/c/jvm/jvm.c
@@ -303,9 +303,7 @@ s32 jvm_init(MiniJVM *jvm, c8 *p_bootclasspath, c8 *p_classpath) {
     reg_net_native_lib(jvm);
     reg_reflect_native_lib(jvm);
     reg_lwjgl_native_lib(jvm);
-    reg_lwjgl_keyboard_native_lib(jvm);
-    reg_lwjgl_mouse_native_lib(jvm);
-    reg_lwjgl_display_native_lib(jvm);
+    reg_lwjgl_platform_native_lib(jvm);
     reg_awt_native_lib(jvm);
 
     //创建jstring 相关容器
